[refactor] 기존 MemberResponseDTO를 MemberDto의 inner class인 MemberResponse로 변경, static 클래스 활용
