+++ src/main/java/springJpaBoard/Board/controller/HomeController.java
@@ -4,10 +4,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import springJpaBoard.Board.controller.responsedto.MemberResponseDTO;
 import springJpaBoard.Board.domain.Member;
 import springJpaBoard.Board.domain.argumenresolver.Login;
 import springJpaBoard.Board.repository.MemberRepository;
+
+import static springJpaBoard.Board.controller.memberdto.MemberDto.MemberResponse;
 
 @Controller
 @RequiredArgsConstructor
@@ -25,7 +26,7 @@ public class HomeController {
 
         /*DTO로 감싸기!*/
         //세션이 유지되면 로그인 홈으로 이동
-        MemberResponseDTO memberResponseDto = new MemberResponseDTO(loginMember);
+        MemberResponse memberResponseDto = MemberResponse.of(loginMember);
 
         model.addAttribute("member", memberResponseDto);
         return "loginHome";
