+++ src/main/java/springJpaBoard/Board/controller/checkInterface/UpdateCheck.java
@@ -1,0 +1,4 @@
+package springJpaBoard.Board.controller.checkInterface;
+
+public interface UpdateCheck {
+}
