+++ src/main/java/springJpaBoard/Board/controller/requestdto/BoardRequestDTO.java
@@ -2,8 +2,8 @@ package springJpaBoard.Board.controller.requestdto;
 
 import lombok.Getter;
 import lombok.Setter;
-import springJpaBoard.Board.controller.requestdto.checkInterface.SaveCheck;
-import springJpaBoard.Board.controller.requestdto.checkInterface.UpdateCheck;
+import springJpaBoard.Board.controller.checkInterface.SaveCheck;
+import springJpaBoard.Board.controller.checkInterface.UpdateCheck;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
