--- src/main/java/springJpaBoard/Board/controller/requestdto/checkInterface/SaveCheck.java
@@ -1,4 +1,0 @@
-package springJpaBoard.Board.controller.requestdto.checkInterface;
-
-public interface SaveCheck {
-}
