--- src/main/java/springJpaBoard/Board/controller/requestdto/checkInterface/UpdateCheck.java
@@ -1,4 +1,0 @@
-package springJpaBoard.Board.controller.requestdto.checkInterface;
-
-public interface UpdateCheck {
-}
